bibstyle_JSSextra {rbibutils}	R Documentation

Create bibstyle JSSextra

Description

Get or create bibstyle JSSextra or clones of it for derived bib styles.

Usage

bibstyle_JSSextra(style = "JSSextra", reset = FALSE, make_default = FALSE)

Arguments

style
the name of the style. The default is "JSSextra".

reset
if TRUE, create a fresh copy of the style even if it exists.

make_default
if TRUE, make the style default.

Details

bibstyle_JSSextra() returns the bib style "JSSextra" (i.e., its environment, not the name of the style). If "JSSextra" has not been created yet, it creates it first.

If argument style is used, as in bibstyle_JSSextra("mystyle"), a style "mystyle" initially equivalent to "JSSextra" is returned (it is created if necessary). The returned environment can be amended to derive a new bib style.

Value

an environment

Author(s)

Georgi N. Boshnakov

[Package rbibutils version 2.4.1 Index]