Convert Between Bibliography Formats

Converts between a number of bibliography formats, including BibTeX, BibLaTeX and Bibentry. Includes a port of the 'bibutils' utilities by Chris Putnam <https://sourceforge.net/projects/bibutils/>.

Details

The DESCRIPTION file: This package was not yet installed at build time.
Index: This package was not yet installed at build time.

Package rbibutils provides an R port of the bibutils programs and functions to convert to R's bibentry Bibtex-based bibliography format.

The core functionality is provided by the bibutils programs which convert between various bibliography formats using a common intermediate MODS XML format, see the source cited below. The bibutils library is included in a number of software packages, including pandoc. Executable programs for conversion are available for Linux distributions but are not easily available for Windows. Executable and libraries can also be generated out-of-the-box from the bibutils disribution (on Windows under MSYS).

Currently we provide the function bibConvert for conversion between supported bibliography formats. For complete list of formats supported by the package, see the documentation of the original bibutils library.

Further functiionality may be provided in future releases, in particular, the underlying C functions could be exposed to package authors. Further R wrappers may be added, as well.

This package will be kept lean. The only external dependency is package xml2 and it is necessary only for conversion to Bibentry and xml manipulation, so could be moved to the suggested packages in a production release. Similar notes hold for util and tools.

References

Chris Putnam, Library bibutils, https://sourceforge.net/projects/bibutils/.