rbibutils_formats {rbibutils}	R Documentation

Supported bibliography formats

Description

Supported bibliography formats in package rbibutils.

Usage

rbibutils_formats

Format

A data frame with 16 observations on the following 5 variables:

Abbreviation

a character vector.

FileExt

a character vector.

Input

a logical vector.

Output

a logical vector.

Description

a character vector.

Details

Each row in rbibutils_formats gives information about a supported bibliography format in package rbibutils.

Abbreviation is the name to use in arguments informat and outformat in bibConvert.

FileExt is the default extension used to infer the input/output format if the corresponding argument is missing.

Input (Output) is TRUE if the format is supported for input (output).

Description gives a brief description of the format.

See bibConvert, section “Supported formats”, for further details.

Examples

rbibutils_formats

[Package rbibutils version 2.4.1 Index]