Convert Between Bibliography Formats

Converts between a number of bibliography formats, including BibTeX, BibLaTeX and Bibentry. Includes a port of the 'bibutils' utilities by Chris Putnam <https://sourceforge.net/projects/bibutils/>. Supports all bibliography formats and character encodings implemented in 'bibutils'.

Details

Package rbibutils provides an R port of the bibutils programs. The main function, bibConvert, offers all conversions between bibliography formats supported by library bibutils. In addition, package rbibutils converts to and from R's bibentry Bibtex-based bibliography format.

The core functionality is provided by the bibutils programs which convert between various bibliography formats using a common MODS XML intermediate format, see the source cited below.

Currently we provide the function bibConvert for conversion between supported bibliography formats. For complete list of formats supported by the package, see the documentation of the original bibutils library.

Further functionality may be provided in future releases, in particular, the underlying C functions could be exposed to package authors. Further R wrappers may be added, as well. However, the scope of the package will remain conversion between formats based on bibutils. rbibutils can be used as an alternative to package bibtex (Francois 2020). For bibliography management, see package RefManager (McLean 2017). For citations in R documentation (Rd or roxygen2) see package Rdpack (Boshnakov 2020).

The only external dependency (import) is package xml2 and conversion between formats not involving bibentry do not need it, so it could be made “suggested” if there is a strong use case for that.

Note

The bibutils library is included in a number of software packages. Thesinclude include pandocand a library for Haskell. Executable programs for conversion are available for Linux distributions but seem not easily available for Windows. Executable and libraries can also be generated out-of-the-box from the bibutils disribution (on Windows under MSYS).

References

Boshnakov GN (2020). “Rdpack: Update and Manipulate Rd Documentation Objects.” doi: 10.5281/zenodo.3925612 (URL: https://doi.org/10.5281/zenodo.3925612), R package version 1.0.0.

Francois R. (2020). bibtex: Bibtex Parser. R package version 0.4.2.2. https://CRAN.R-project.org/package=bibtex.

McLean MW (2017). “RefManageR: Import and Manage BibTeX and BibLaTeX References in R.” _The Journal of Open Source Software_. doi: 10.21105/joss.00338 (URL: https://doi.org/10.21105/joss.00338).

Putnam C. (2003--2020), Library bibutils, https://sourceforge.net/projects/bibutils/.

See also