Import bibtex files

Import bibtex files.

readBib(file, encoding)

writeBib(object, con = stdout(), append = FALSE)

Arguments

file
name or path to the file, a character string.

encoding
the encoding of file, a character string.

object

con
filename (a character string) or a text connection

append
if TRUE append to the file.

Details

Read a bibtex file and convert the result to bibentry or write a bibentry to a bibtex file.

Value

for readBib, a bibentry object

for writeBib, the bibentry object (invisibly)

See also

Examples

## create a bibentry object
bibs <- readBib(system.file("REFERENCES.bib", package = "rbibutils"),
                encoding = "UTF-8")
## write bibs to a file
fn <- tempfile(fileext = ".bib")
writeBib(bibs, fn)

## see the contents of the file
readLines(fn) # or: file.show(fn)

#>  [1] "@Manual{Rpackage:bibtex,"                                                      
#>  [2] "  author = {Romain Francois},"                                                 
#>  [3] "  title = {bibtex: bibtex parser},"                                            
#>  [4] "  year = {2014},"                                                              
#>  [5] "  note = {R package version 0.4.0},"                                           
#>  [6] "}"                                                                             
#>  [7] ""                                                                              
#>  [8] "@Misc{Rpackage:Rdpack,"                                                        
#>  [9] "  author = {Georgi N Boshnakov},"                                              
#> [10] "  title = {{Rdpack}: Update and Manipulate Rd Documentation Objects},"         
#> [11] "  year = {2020},"                                                              
#> [12] "  note = {R package version 2.0.0},"                                           
#> [13] "  doi = {10.5281/zenodo.3925612},"                                             
#> [14] "  url = {https://doi.org/10.5281/zenodo.3925612},"                             
#> [15] "}"                                                                             
#> [16] ""                                                                              
#> [17] "@Article{Rpackage:RefManageR,"                                                 
#> [18] "  author = {Mathew William McLean},"                                           
#> [19] "  title = {RefManageR: Import and Manage BibTeX and BibLaTeX References in R},"
#> [20] "  journal = {The Journal of Open Source Software},"                            
#> [21] "  year = {2017},"                                                              
#> [22] "  doi = {10.21105/joss.00338},"                                                
#> [23] "  url = {https://doi.org/10.21105/joss.00338},"                                
#> [24] "}"                                                                             

unlink(fn)