Read and write bibentry files

Read and write bibentry files.

readBibentry(file)

writeBibentry(be, file, style = c("Rstyle", "loose"))

Arguments

be
a bibentry object.

file
filename, a character string.

style
if "Rstyle" (default), wrap in c(), otherwise don't wrap and don't put commas between the entries, see Details section.

Details

These functions read/write bibentry objects from/to R source files. Two styles are supported. "Rstyle" is the format used by print(be, style = "R"), which writes the bibentry calls as a comma separated sequence wrapped in c(). Style "loose" writes the entries without separators and no wrapping.

writeBibentry writes the object to the specified file in the requested style (default is "Rstyle").

readBibentry reads the file and creates a bibentry object. It doesn't have argument for style, since that is inferred from the contents of the file.

bibentry() calls that throw errors are not included in the returned object. The errors are intercepted and converted to warnings, identifying the corresponding bibentry() calls by their keys, if present (otherwise the text of the whole bibentry is shown).

Value

for writeBibentry, NULL (invisibly)

for readBibentry, a bibentry object with the keys as names

See also

Examples